I'm putting the project on hiatus. I have lost some motivation, and been inspired by simpler projects. I hope to return to this at a later date with fresh eyes, and a better grasp of the complexities involved in everything I'm trying to do.
This last commit is simply a snapshot of where I was before I decided to
jump ship. No cleanup. It may be necescary to jump back a few commits in
order to have a clean slate, but I thought it best to make a record of
where I stopped. Also I'm lazy. The last things I was working on were
getting the prey to bulge using normal maps, and rotate by either
modifying the curves for distortion, or actually importing the mesh
entirely and just rotating it in game. Anyways, that's a problem for
future me to figure out.